+++ main.js
@@ -11,91 +11,115 @@
 
 "use strict";
 
-let {
+const {
 	commands,
 	window
 } = require('vscode');
-let FileSaver = require('fs');
-let vscode = require('vscode');
-let HexMinifier = require('./Minifiers/hexMinifier');
-let LineRemover = require('./Minifiers/lineRemover');
+const FileSaver = require('fs');
+const vscode = require('vscode');
+const HexMinifier = require('./src/hexMinifier');
+const LineRemover = require('./src/lineRemover');
 let originalFilepath = vscode.window.activeTextEditor.document.fileName
 let originalSize = FileSaver.statSync(originalFilepath).size
+const ModuleSizeTransform = require('./src/sizeTransform');
+const sizeTransform = new ModuleSizeTransform();
 
 vscode.commands.registerCommand('extension.MinifyAllStatus', statusBarInfo);
 vscode.workspace.onDidSaveTextDocument(() => getNewSize());
+const statusBarItem = vscode.window.createStatusBarItem(vscode.StatusBarAlignment.Left, 1);
 
+/**
+ * getNewSize gets the new size of the
+ * document and creates the triggers
+ * then calls createStatusBar()
+ */
 function getNewSize() {
-	let newFilepath = vscode.window.activeTextEditor.document.fileName
-	let newSize = FileSaver.statSync(newFilepath).size
+	const newFilepath = vscode.window.activeTextEditor.document.fileName
+	const newSize = FileSaver.statSync(newFilepath).size
 	createStatusBar(originalSize, newSize);
+	vscode.workspace.onDidChangeConfiguration(() => statusBarItem.hide());
+	vscode.workspace.onDidChangeWorkspaceFolders(() => statusBarItem.hide());
+	vscode.workspace.onDidCloseTextDocument(() => statusBarItem.hide());
 }
 
+/**
+ * createStatusBar creates the status bar item
+ * @param {number} originalSize the unminified size
+ * @param {number} newSize the minified size
+ */
 function createStatusBar(originalSize, newSize) {
-	let statusBarItem = vscode.window.createStatusBarItem(vscode.StatusBarAlignment.Left, 1);
 	statusBarItem.tooltip = 'New file size, click for more info!';
 	statusBarItem.command = 'extension.MinifyAllStatus';
-	statusBarItem.text = transformSize(originalSize) + " --> " + transformSize(newSize);
+	statusBarItem.text = sizeTransform.transformSize(originalSize) + " --> " + sizeTransform.transformSize(newSize);
 	statusBarItem.show();
 	vscode.workspace.onDidChangeConfiguration(() => statusBarItem.hide());
 	vscode.workspace.onDidChangeWorkspaceFolders(() => statusBarItem.hide());
 	vscode.workspace.onDidCloseTextDocument(() => statusBarItem.hide());
 }
 
+/**
+ * statusBarInfo Creates an output
+ * channel with information about
+ * the file that has been minified
+ * original size, new minified size
+ * filetype and path.
+ */
 function statusBarInfo() {
-	let oc = window.createOutputChannel('Minify output');
+	const oc = window.createOutputChannel('Minify output');
 	oc.appendLine("╔══════════════════════════════╗");
 	oc.appendLine("║      Extension MinifyAll     ║	")
 	oc.appendLine("╠═══════════════════╦══════════╣");
-	oc.appendLine("║ Original size     ║ " + transformSize(originalSize) + " ║");
+	oc.appendLine("║ Original size     ║ " + sizeTransform.transformSize(originalSize) + "  ║");
 	oc.appendLine("╠═══════════════════╬══════════║");
-	oc.appendLine("║ New minified size ║ " + transformSize(FileSaver.statSync(vscode.window.activeTextEditor.document.fileName).size) + " ║");
+	oc.appendLine("║ New minified size ║ " + sizeTransform.transformSize(FileSaver.statSync(vscode.window.activeTextEditor.document.fileName).size) + "  ║");
 	oc.appendLine("╚═══════════════════╩══════════╝");
 	oc.appendLine("File path:\t" + window.activeTextEditor.document.fileName);
+	oc.appendLine("File type:\t" + window.activeTextEditor.document.languageId);
 	oc.show();
 }
 
-function transformSize(size) {
-	if (size >= 1048576) return `${Math.floor(size / 10485.76) / 100} MB`;
-	else if (size >= 1024) return `${Math.floor(size / 10.24) / 100} KB`;
-	else return `${size} B`;
-}
-
+/**
+ * activate Main function called when the user
+ * uses the command 'MinifyAll' or 'MinifyAll2OtherDoc'
+ * @param {object} context 
+ */
 function activate(context) {
 	//Command MinifyAll default one.
-	let disposable = commands.registerCommand('extension.MinifyAll', () => {
-		let {
+	const disposable = commands.registerCommand('extension.MinifyAll', () => {
+		originalFilepath = vscode.window.activeTextEditor.document.fileName
+		originalSize = FileSaver.statSync(originalFilepath).size
+		const {
 			document
 		} = window.activeTextEditor;
 		switch (window.activeTextEditor.document.languageId) {
 			case "css":
 			case "scss":
 
-				let editorCss = vscode.window.activeTextEditor;
-				let firstLineCss = editorCss.document.lineAt(0);
-				let lastLineCss = editorCss.document.lineAt(editorCss.document.lineCount - 1);
-				let textRangeCss = new vscode.Range(0,
+				const editorCss = vscode.window.activeTextEditor;
+				const firstLineCss = editorCss.document.lineAt(0);
+				const lastLineCss = editorCss.document.lineAt(editorCss.document.lineCount - 1);
+				const textRangeCss = new vscode.Range(0,
 					firstLineCss.range.start.character,
 					editorCss.document.lineCount - 1,
 					lastLineCss.range.end.character);
 
-				let cssMinifier = require('./Minifiers/cssMinifier.js');
-				let cssContent = document.getText().split('\n');
-				let RemoverLine4Css = new LineRemover(cssContent);
+				const cssMinifier = require('./src/cssMinifier.js');
+				const cssContent = document.getText().split('\n');
+				const RemoverLine4Css = new LineRemover(cssContent);
 				RemoverLine4Css.removeMultipleLineComments();
 				RemoverLine4Css.removeSingleLineComments();
 
-				let MinifierHex4Css = new HexMinifier(RemoverLine4Css.getLineRemoved());
+				const MinifierHex4Css = new HexMinifier(RemoverLine4Css.getLineRemoved());
 
 				//Minifier processes
 				MinifierHex4Css.shortHexMain();
 				MinifierHex4Css.shortRGBMain();
 				MinifierHex4Css.shortRGBAMain();
 
-				let minifiercss = new cssMinifier(MinifierHex4Css.getHexMinified());
+				const minifiercss = new cssMinifier(MinifierHex4Css.getHexMinified());
 
 				//Get the minified code and replace it
-				let modifiedCssText = minifiercss.getCssMinified();
+				const modifiedCssText = minifiercss.getCssMinified();
 				editorCss.edit(builder => {
 					builder.replace(textRangeCss, modifiedCssText);
 				});
@@ -104,32 +128,33 @@ function activate(context) {
 
 			case "json":
 			case "jsonc":
-				let editorJson = vscode.window.activeTextEditor;
-				let firstLineJson = editorJson.document.lineAt(0);
-				let lastLineJson = editorJson.document.lineAt(editorJson.document.lineCount - 1);
-				let textRangeJson = new vscode.Range(0,
+
+				const editorJson = vscode.window.activeTextEditor;
+				const firstLineJson = editorJson.document.lineAt(0);
+				const lastLineJson = editorJson.document.lineAt(editorJson.document.lineCount - 1);
+				const textRangeJson = new vscode.Range(0,
 					firstLineJson.range.start.character,
 					editorJson.document.lineCount - 1,
 					lastLineJson.range.end.character);
 
-				let jsonMinifier = require('./Minifiers/jsonMinifier.js');
-				let jsonContent = document.getText().split('\n');
-				let MinifierHex4Json = new HexMinifier(jsonContent);
+				const jsonMinifier = require('./src/jsonMinifier.js');
+				const jsonContent = document.getText().split('\n');
+				const MinifierHex4Json = new HexMinifier(jsonContent);
 
 				//Minifier processes
 				MinifierHex4Json.shortHexMain();
 				MinifierHex4Json.shortRGBMain();
 				MinifierHex4Json.shortRGBAMain();
 
-				let RemoverLine4Json = new LineRemover(MinifierHex4Json.getHexMinified());
+				const RemoverLine4Json = new LineRemover(MinifierHex4Json.getHexMinified());
 
 				RemoverLine4Json.removeMultipleLineComments();
 				RemoverLine4Json.removeSingleLineComments();
 
-				let minifierjson = new jsonMinifier(RemoverLine4Json.getLineRemoved());
+				const minifierjson = new jsonMinifier(RemoverLine4Json.getLineRemoved());
 
 				//Get the minified code and replace it
-				let modifiedJsonText = minifierjson.getJSONMinified();
+				const modifiedJsonText = minifierjson.getJSONMinified();
 
 				editorJson.edit(builder => {
 					builder.replace(textRangeJson, modifiedJsonText);
@@ -137,21 +162,21 @@ function activate(context) {
 				break;
 
 			case "html":
-				let editorHtml = vscode.window.activeTextEditor;
 
-				let firstLineHtml = editorHtml.document.lineAt(0);
-				let lastLineHtml = editorHtml.document.lineAt(editorHtml.document.lineCount - 1);
-				let textRangeHtml = new vscode.Range(0,
+				const editorHtml = vscode.window.activeTextEditor;
+				const firstLineHtml = editorHtml.document.lineAt(0);
+				const lastLineHtml = editorHtml.document.lineAt(editorHtml.document.lineCount - 1);
+				const textRangeHtml = new vscode.Range(0,
 					firstLineHtml.range.start.character,
 					editorHtml.document.lineCount - 1,
 					lastLineHtml.range.end.character);
 
-				let htmlMinifier = require('./Minifiers/htmlMinifier.js');
-				let htmlContent = document.getText().split('\n');
-				let minifierhtml = new htmlMinifier(htmlContent);
+				const htmlMinifier = require('./src/htmlMinifier.js');
+				const htmlContent = document.getText().split('\n');
+				const minifierhtml = new htmlMinifier(htmlContent);
 				minifierhtml.removeMultipleLineComments();
 				//Get the minified code and replace it
-				let modifiedHtmlText = minifierhtml.gethtmlMinified();
+				const modifiedHtmlText = minifierhtml.gethtmlMinified();
 
 				editorHtml.edit(builder => {
 					builder.replace(textRangeHtml, modifiedHtmlText);
@@ -168,39 +193,39 @@ function activate(context) {
 	});
 
 	//Command MinifyAll and writes the result in other file.
-	let disposable2 = commands.registerCommand('extension.MinifyAll2OtherDoc', () => {
-		let path = require('path');
-		let {
+	const disposable2 = commands.registerCommand('extension.MinifyAll2OtherDoc', () => {
+		const path = require('path');
+		const {
 			document
 		} = window.activeTextEditor;
-		let {
+		const {
 			fileName
 		} = document;
 
-		let filePath = path.dirname(fileName);
+		const filePath = path.dirname(fileName);
 		switch (window.activeTextEditor.document.languageId) {
 
 			case "css":
 			case "scss":
-				let newName = path.basename(fileName).replace('.css', '-min.css');
-				let path2NewFile = path.join(filePath, newName);
-				let cssMinifier = require('./Minifiers/cssMinifier.js');
-				let cssContent = document.getText().split('\n');
-				let RemoverLine4Css = new LineRemover(cssContent);
+				const newName = path.basename(fileName).replace('.css', '-min.css');
+				const path2NewFile = path.join(filePath, newName);
+				const cssMinifier = require('./src/cssMinifier.js');
+				const cssContent = document.getText().split('\n');
+				const RemoverLine4Css = new LineRemover(cssContent);
 
 				RemoverLine4Css.removeMultipleLineComments();
 				RemoverLine4Css.removeSingleLineComments();
 
-				let MinifierHex4Css = new HexMinifier(RemoverLine4Css.getLineRemoved());
+				const MinifierHex4Css = new HexMinifier(RemoverLine4Css.getLineRemoved());
 
 				//Minifier processes
 				MinifierHex4Css.shortHexMain();
 				MinifierHex4Css.shortRGBMain();
 				MinifierHex4Css.shortRGBAMain();
 
-				let minifiercss = new cssMinifier(MinifierHex4Css.getHexMinified());
+				const minifiercss = new cssMinifier(MinifierHex4Css.getHexMinified());
 
-				let modifiedCssText = minifiercss.getCssMinified();
+				const modifiedCssText = minifiercss.getCssMinified();
 
 				//Get the minified code and replace it in a new file
 				FileSaver.writeFile(path2NewFile, modifiedCssText, () => {
@@ -215,26 +240,26 @@ function activate(context) {
 			case "json":
 			case "jsonc":
 
-				let newNamejson = path.basename(fileName).replace('.json', '-min.json');
-				let path2NewFilejson = path.join(filePath, newNamejson);
-				let jsonMinifier = require('./Minifiers/jsonMinifier.js');
-				let jsonContent = document.getText().split('\n');
-				let MinifierHex4Json = new HexMinifier(jsonContent);
+				const newNamejson = path.basename(fileName).replace('.json', '-min.json');
+				const path2NewFilejson = path.join(filePath, newNamejson);
+				const jsonMinifier = require('./src/jsonMinifier.js');
+				const jsonContent = document.getText().split('\n');
+				const MinifierHex4Json = new HexMinifier(jsonContent);
 
 				//Minifier processes
 				MinifierHex4Json.shortHexMain();
 				MinifierHex4Json.shortRGBMain();
 				MinifierHex4Json.shortRGBAMain();
 
-				let RemoverLine4Json = new LineRemover(MinifierHex4Json.getHexMinified());
+				const RemoverLine4Json = new LineRemover(MinifierHex4Json.getHexMinified());
 
 				RemoverLine4Json.removeMultipleLineComments();
 				RemoverLine4Json.removeSingleLineComments();
 
-				let minifierjson = new jsonMinifier(RemoverLine4Json.getLineRemoved());
+				const minifierjson = new jsonMinifier(RemoverLine4Json.getLineRemoved());
 
 				//Get the minified code and replace it
-				let modifiedJsonText = minifierjson.getJSONMinified();
+				const modifiedJsonText = minifierjson.getJSONMinified();
 
 				//Get the minified code and replace it in a new file
 				FileSaver.writeFile(path2NewFilejson, modifiedJsonText, () => {
@@ -248,14 +273,14 @@ function activate(context) {
 
 			case "html":
 
-				let newNamehtml = path.basename(fileName).replace('.html', '-min.html');
-				let path2NewFilehtml = path.join(filePath, newNamehtml);
-				let htmlMinifier = require('./Minifiers/htmlMinifier.js');
-				let htmlContent = document.getText().split('\n');
-				let minifierhtml = new htmlMinifier(htmlContent);
+				const newNamehtml = path.basename(fileName).replace('.html', '-min.html');
+				const path2NewFilehtml = path.join(filePath, newNamehtml);
+				const htmlMinifier = require('./src/htmlMinifier.js');
+				const htmlContent = document.getText().split('\n');
+				const minifierhtml = new htmlMinifier(htmlContent);
 				minifierhtml.removeMultipleLineComments();
 
-				let modifiedHtmlText = minifierhtml.gethtmlMinified();
+				const modifiedHtmlText = minifierhtml.gethtmlMinified();
 
 				//Get the minified code and replace it in a new file
 				FileSaver.writeFile(path2NewFilehtml, modifiedHtmlText, () => {
@@ -280,5 +305,12 @@ function activate(context) {
 
 exports.activate = activate;
 
-function deactivate() {}
+/**
+ * deactivate function that is
+ * called when the extension
+ * is deactivaded.
+ */
+function deactivate() {
+	statusBarItem.dispose();
+}
 exports.deactivate = deactivate;